updating the FAQ page to include NavigationBar
