+++ client/src/App.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import Home from "./pages/Home";
 import LearningMaterialSummary from "./pages/LearningMaterialSummary";
 import About from "./pages/About";
+import Questions from "./pages/Questions";
 
 const App: React.FC<{}> = () => {
   return (
@@ -11,6 +12,7 @@ const App: React.FC<{}> = () => {
         <Routes>
           <Route path="/" element={<Home />} />
           <Route path="/learningMaterialSummary" element={<LearningMaterialSummary />} />
+          <Route path="/faq" element={<Questions />} />
           <Route path="/about" element={<About />} />
         </Routes>
       </div>
